modify contact, about, header
Added footer/headers for contact, about pages
Improved contact us structure
refactored logout, authentication (logged in) logic to header from home
