+++ github-project/src/App.css
@@ -579,24 +579,24 @@ body.dark-mode {
     color: var(--white);
 }
 
-.contact-page {
+.contact-container {
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding: 40px 20px;
   min-height: 80vh;
-  background-color: #dad7cd;
+  background-color: inherit;
 }
 
 .contact-form {
+  margin-top: 20vh;
   display: flex;
   flex-direction: column;
   gap: 15px;
   background-color: #a3b18a;
   padding: 30px;
   border-radius: 10px;
-  width: 100%;
-  max-width: 500px;
+  width: 50vw;
+  max-width: 750px;
   color: #344e41;
 }
